+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Nothing
 
 ### Fixed
-- Nothing
+- Capitalizer discarded empty strings.
 
 ### Deprecated
 - Nothing